+++ mathlib/src/objects/material.rs
@@ -4,10 +4,10 @@ use crate::{cmp::ApproxEq, visual::color::Col};
 #[derive(Debug, Clone)]
 pub struct Material {
     pub color: Col,
-    pub ambient: f32,
-    pub diffuse: f32,
-    pub specular: f32,
-    pub shininess: f32,
+    pub ambient: f64,
+    pub diffuse: f64,
+    pub specular: f64,
+    pub shininess: f64,
 }
 
 impl Material {
