+++ runtime_unix/src/main.rs
@@ -19,12 +19,12 @@ fn main() {
 
 pub fn manually_cast_rays_at_sphere_infront_canvas() {
     let ray_origin = Point::inew(0, 0, -5);
-    let wall_z: f32 = 10.0;
+    let wall_z: f64 = 10.0;
     let wall_size = 7.0;
 
     let canvas_size_px = 100;
     let mut canvas = Canvas::new(canvas_size_px, canvas_size_px);
-    let pixel_size = wall_size / canvas_size_px as f32;
+    let pixel_size = wall_size / canvas_size_px as f64;
     let half = wall_size / 2.0;
 
     let mut shape = Sphere::new();
@@ -40,11 +40,11 @@ pub fn manually_cast_rays_at_sphere_infront_canvas() {
     // for each pixel:
     for (y, row) in canvas.arr.iter_mut().enumerate() {
         // compute current world y choordinate
-        let world_y = half - pixel_size * y as f32;
+        let world_y = half - pixel_size * y as f64;
 
         for (x, col) in row.iter_mut().enumerate() {
             // compute the world x choordinate
-            let world_x = -half + pixel_size * x as f32;
+            let world_x = -half + pixel_size * x as f64;
             // Point the ray will hit:
             let position = Point::new(world_x, world_y, wall_z);
 
@@ -79,7 +79,7 @@ pub fn manually_create_gradient_file() {
     let mut canvas = Canvas::new(100, 200);
     for (i, row) in canvas.arr.iter_mut().enumerate() {
         for (j, col) in row.iter_mut().enumerate() {
-            *col = Col::new(i as f32 / 100.0, 0.8, j as f32 / 200.0);
+            *col = Col::new(i as f64 / 100.0, 0.8, j as f64 / 200.0);
         }
     }
     write_to_file("./out.ppm", canvas.canvas_to_ppm());
